requirements.txt
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.14.0
pandas>=2.0.0
plotly>=5.15.0
requests>=2.31.0
pymongo>=4.5.0
python-dotenv>=1.0.0
geopy>=2.3.0
openai>=1.0.0
datetime
pytz>=2023.3